refactor: move runtime code to `crate::rt`
