+++ src/arch/aarch64/start/hermit_entry.rs
@@ -92,7 +92,7 @@ pub unsafe extern "C" fn pre_init(boot_info: Option<&'static RawBootInfo>, cpu_i
 
 	if cpu_id == 0 {
 		env::set_boot_info(*boot_info.unwrap());
-		crate::boot_processor_main()
+		crate::rt::boot_processor_main()
 	} else {
 		#[cfg(not(feature = "smp"))]
 		{
@@ -106,6 +106,6 @@ pub unsafe extern "C" fn pre_init(boot_info: Option<&'static RawBootInfo>, cpu_i
 			}
 		}
 		#[cfg(feature = "smp")]
-		crate::application_processor_main()
+		crate::rt::application_processor_main()
 	}
 }
