+++ src/arch/x86_64/start/hermit_entry.rs
@@ -78,7 +78,7 @@ unsafe extern "C" fn pre_init(boot_info: Option<&'static RawBootInfo>, cpu_id: u
 	if cpu_id == 0 {
 		env::set_boot_info(*boot_info.unwrap());
 
-		crate::boot_processor_main()
+		crate::rt::boot_processor_main()
 	} else {
 		#[cfg(not(feature = "smp"))]
 		{
@@ -92,6 +92,6 @@ unsafe extern "C" fn pre_init(boot_info: Option<&'static RawBootInfo>, cpu_id: u
 			}
 		}
 		#[cfg(feature = "smp")]
-		crate::application_processor_main();
+		crate::rt::application_processor_main();
 	}
 }
