+++ src/arch/x86_64/start/smp.rs
@@ -29,5 +29,5 @@ unsafe extern "C" fn smp_start_rust() -> ! {
 		Cr0::update(|flags| flags.remove(Cr0Flags::CACHE_DISABLE | Cr0Flags::NOT_WRITE_THROUGH));
 	}
 
-	crate::application_processor_main();
+	crate::rt::application_processor_main();
 }
